pokemonCardDocker/requirementsLambda.txt
# Editable install with no version control (DatabaseImageMatching==1.0.0)
boto3
numpy==1.25.1
#opencv-python==4.8.0.74
opencv-python-headless
Pillow==10.0.0
